llvm-gcc side of arm eh support
First half of arm eh support.
This pretty much only changes two register numbers.
And adds a bunch of comments.

There is an undocumented contract between gcc and the unwinder that the
registers used to communicate handler information must be callee saved regs in
the unwinding routine, e.g. _Unwind_RaiseException.

As far as I can tell this is not enforced in anyway but instead there is a
runtime check.
So we are relying on luck here.
This is true for x86 as well.

Luckily _Unwind_RaiseException is complicated enough that lots of registers are
saved  in its prolog.

Review at http://codereview.chromium.org/5903004/